Removed HL7Parser and moved message parsing logic to HL7Message to avoid circular dependencies
